Departements tests done and saveReservation test done

+++ src/main/java/com/youcode/resourcium/Entities/Departement.java
@@ -10,7 +10,7 @@ public class Departement {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -23,7 +23,7 @@ public class Departement {
     private User ChefDepartment;
 
     public Departement(){}
-    public Departement(int id, String name, String description) {
+    public Departement(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -41,11 +41,11 @@ public class Departement {
         this.ChefDepartment = Chef;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
